requirements-dev.txt
ruff==0.15.4
pytest==9.0.2
jsonschema==4.26.0
fastapi==0.115.12
httpx==0.28.1
uvicorn==0.34.3
pydantic==2.11.7
PyJWT==2.7.0
